feat: yield curve predictor, liquidation dashboard, liquidation bot, and opportunity explorer
- Yield curve prediction model with 1d/7d/30d horizons, confidence intervals, MAPE tracking
- Liquidation monitoring dashboard with WS real-time health updates and configurable alerts
- Automated liquidation bot with configurable profitability thresholds and dry-run mode
- Liquidation opportunity explorer with gas estimation, filtering, and sorting
- Smart contract yield curve module (PiecewiseLinear, Polynomial, Nelson-Siegel)
- Frontend React components for all dashboards
- Full test coverage: 351 tests passing

+++ stellar-lend/contracts/lending/src/storage.rs
@@ -82,7 +82,7 @@ pub struct PoolConfig {
     pub liquidation_incentive_bps: i128,
     pub last_update: u64,
     /// Packed status-flag byte (see `FLAG_*`).
-    pub flags: u8,
+    pub flags: u32,
 }
 
 /// The two packed words as persisted on-chain.
@@ -135,7 +135,7 @@ pub fn unpack(packed: &PackedConfig) -> PoolConfig {
         close_factor_bps: unpack_bps_field(packed.rate_word, 3),
         liquidation_incentive_bps: unpack_bps_field(packed.rate_word, 4),
         last_update: packed.status_word & TS_MASK,
-        flags: ((packed.status_word >> FLAGS_SHIFT) & FLAGS_MASK) as u8,
+        flags: ((packed.status_word >> FLAGS_SHIFT) & FLAGS_MASK) as u32,
     }
 }
